Move units_id to common in MELCloud Home
